# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Adaptive Psychological Monitoring Platform",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/interviews/start": {
      "post": {
        "tags": [
          "Adaptive Interview"
        ],
        "summary": "Start Interview",
        "description": "Start a new Adaptive Interview session.\n\nThe Adaptive Interview Engine selects the first question based on the\npatient's chronic condition(s) (Disease Knowledge Layer priority\ncategories) and returns it as `bot_message_ar`.\n\nFails with `409 Conflict` if the patient already has an in-progress\nsession - it must be completed via `/answer` or `/end` first.\n\nIn the rare case where no questions are configured at all, the session\nends immediately (`is_session_ended=true`) and a risk assessment is still\nproduced from the (empty) session.",
        "operationId": "start_interview_api_v1_interviews_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewTurnResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/interviews/{session_id}/answer": {
      "post": {
        "tags": [
          "Adaptive Interview"
        ],
        "summary": "Submit Answer",
        "description": "Submit the patient's answer to the current question and advance the\ninterview.\n\nSend `answer_text_ar` for open-ended questions (the Sentiment Analysis\nEngine classifies the emotion: anxiety/stress/sadness/burnout/frustration/\npositive/neutral) or `answer_value_numeric` for `scale_1_5` /\n`yes_no` / `multiple_choice` questions.\n\nBehavior:\n- If the detected emotion strongly matches a Disease Knowledge Layer\n  pattern for the patient's condition(s), the Question Selector\n  **escalates** - the next question probes deeper into the related\n  category.\n- If the answer contains explicit crisis/self-harm language, the session\n  ends immediately (`is_session_ended=true`), `risk_assessment_id` points\n  to a **Level 5** assessment, a `professional_help` recommendation is\n  created, and admins receive a `risk_alert_admin` notification.\n- Otherwise, when the Depth/Termination Controller decides the session is\n  complete (minimum questions reached and all priority categories\n  covered, or the maximum question count is hit), `is_session_ended=true`\n  and `risk_assessment_id` is populated.\n- While `is_session_ended=false`, `bot_message_ar` contains the next\n  question.",
        "operationId": "submit_answer_api_v1_interviews__session_id__answer_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewAnswerSubmit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewTurnResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{session_id}/end": {
      "post": {
        "tags": [
          "Adaptive Interview"
        ],
        "summary": "End Interview",
        "description": "End an in-progress interview session early (e.g., the patient wants to\nstop before the engine would normally terminate it).\n\nA risk assessment and recommendations are still generated from whatever\nanswers were collected before ending.",
        "operationId": "end_interview_api_v1_interviews__session_id__end_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewTurnResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/history": {
      "get": {
        "tags": [
          "Adaptive Interview"
        ],
        "summary": "Interview History",
        "operationId": "interview_history_api_v1_interviews_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InterviewSessionOut"
                  },
                  "type": "array",
                  "title": "Response Interview History Api V1 Interviews History Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/interviews/{session_id}": {
      "get": {
        "tags": [
          "Adaptive Interview"
        ],
        "summary": "Get Interview",
        "operationId": "get_interview_api_v1_interviews__session_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{session_id}/conversation": {
      "get": {
        "tags": [
          "Adaptive Interview"
        ],
        "summary": "Get Interview Conversation",
        "operationId": "get_interview_conversation_api_v1_interviews__session_id__conversation_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/interviews/{session_id}": {
      "get": {
        "tags": [
          "Admin Dashboard"
        ],
        "summary": "Get Interview Session",
        "operationId": "get_interview_session_api_v1_admin_interviews__session_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminInterviewDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminAnswerOut": {
        "properties": {
          "question_text_ar_snapshot": {
            "type": "string",
            "title": "Question Text Ar Snapshot"
          },
          "answer_text_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Text Ar"
          },
          "answer_value_numeric": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Value Numeric"
          },
          "sentiment_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Label"
          },
          "sentiment_score": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Score"
          },
          "sequence_order": {
            "type": "integer",
            "title": "Sequence Order"
          }
        },
        "type": "object",
        "required": [
          "question_text_ar_snapshot",
          "sequence_order"
        ],
        "title": "AdminAnswerOut"
      },
...
      "AdminInterviewDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Profile Id"
          },
          "patient_full_name": {
            "type": "string",
            "title": "Patient Full Name"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "trigger_type": {
            "$ref": "#/components/schemas/TriggerType"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "total_questions_asked": {
            "type": "integer",
            "title": "Total Questions Asked"
          },
          "session_summary_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Summary Ar"
          },
          "conversation": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageOut"
            },
            "type": "array",
            "title": "Conversation",
            "default": []
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AdminAnswerOut"
            },
            "type": "array",
            "title": "Answers",
            "default": []
          },
          "risk_assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminRiskAssessmentDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_profile_id",
          "patient_full_name",
          "status",
          "trigger_type",
          "started_at",
          "total_questions_asked"
        ],
        "title": "AdminInterviewDetail"
      },
...
      "AdminRiskAssessmentDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "risk_level": {
            "type": "integer",
            "title": "Risk Level"
          },
          "anxiety_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Anxiety Score"
          },
          "stress_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Stress Score"
          },
          "sadness_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Sadness Score"
          },
          "burnout_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Burnout Score"
          },
          "sleep_quality_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Sleep Quality Score"
          },
          "adherence_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Adherence Score"
          },
          "composite_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Composite Score"
          },
          "explanation_ar": {
            "type": "string",
            "title": "Explanation Ar"
          },
          "explanation_factors_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Explanation Factors Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "risk_level",
          "anxiety_score",
          "stress_score",
          "sadness_score",
          "burnout_score",
          "sleep_quality_score",
          "adherence_score",
          "composite_score",
          "explanation_ar",
          "explanation_factors_json",
          "created_at"
        ],
        "title": "AdminRiskAssessmentDetail"
      },
...
      "ChatMessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "sender": {
            "$ref": "#/components/schemas/ChatSender"
          },
          "message_text_ar": {
            "type": "string",
            "title": "Message Text Ar"
          },
          "message_order": {
            "type": "integer",
            "title": "Message Order"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sender",
          "message_text_ar",
          "message_order",
          "created_at"
        ],
        "title": "ChatMessageOut"
      },
      "ChatSender": {
        "type": "string",
        "enum": [
          "bot",
          "patient"
        ],
        "title": "ChatSender"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterviewAnswerSubmit": {
        "properties": {
          "answer_text_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Text Ar"
          },
          "answer_value_numeric": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Value Numeric"
          }
        },
        "type": "object",
        "title": "InterviewAnswerSubmit",
        "description": "Patient's reply to the current question.\n- answer_text_ar: free-text Arabic reply (for open_text questions)\n- answer_value_numeric: numeric reply (for scale_1_5 / yes_no / multiple_choice questions)"
      },
      "InterviewSessionDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "trigger_type": {
            "$ref": "#/components/schemas/TriggerType"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "total_questions_asked": {
            "type": "integer",
            "title": "Total Questions Asked"
          },
          "session_summary_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Summary Ar"
          },
          "conversation": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageOut"
            },
            "type": "array",
            "title": "Conversation",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "trigger_type",
          "started_at",
          "total_questions_asked"
        ],
        "title": "InterviewSessionDetailOut"
      },
      "InterviewSessionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "trigger_type": {
            "$ref": "#/components/schemas/TriggerType"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "total_questions_asked": {
            "type": "integer",
            "title": "Total Questions Asked"
          },
          "session_summary_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Summary Ar"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "trigger_type",
          "started_at",
          "total_questions_asked"
        ],
        "title": "InterviewSessionOut"
      },
      "InterviewStartRequest": {
        "properties": {
          "trigger_type": {
            "$ref": "#/components/schemas/TriggerType",
            "default": "manual"
          }
        },
        "type": "object",
        "title": "InterviewStartRequest"
      },
      "InterviewTurnResponse": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/InterviewSessionOut"
          },
          "bot_message_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bot Message Ar"
          },
          "is_session_ended": {
            "type": "boolean",
            "title": "Is Session Ended",
            "default": false
          },
          "risk_assessment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Assessment Id"
          }
        },
        "type": "object",
        "required": [
          "session"
        ],
        "title": "InterviewTurnResponse",
        "description": "Response returned after starting a session or submitting an answer.\nEither contains the next bot message (question) or signals session end."
      },
...
      "SessionStatus": {
        "type": "string",
        "enum": [
          "in_progress",
          "completed",
          "abandoned"
        ],
        "title": "SessionStatus"
      },
...
      "TriggerType": {
        "type": "string",
        "enum": [
          "daily_checkin",
          "manual",
          "follow_up",
          "scheduled"
        ],
        "title": "TriggerType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}